refactor: extract unit actions into its controller
